Add handler for contract UI
