+++ src/handle_finalize.c
@@ -4,8 +4,6 @@ void handle_finalize(ethPluginFinalize_t *msg) {
     context_t *context = (context_t *) msg->pluginContext;
 
     msg->uiType = ETH_UI_TYPE_GENERIC;
-
-    // EDIT THIS: Set the total number of screen you will need.
     msg->numScreens = 2;
     // EDIT THIS: Handle this case like you wish to (i.e. maybe no additional screen needed?).
     // If the beneficiary is NOT the sender, we will need an additional screen to display it.
@@ -13,9 +11,19 @@ void handle_finalize(ethPluginFinalize_t *msg) {
         msg->numScreens += 1;
     }
 
-    // EDIT THIS: set `tokenLookup1` (and maybe `tokenLookup2`) to point to
-    // token addresses you will info for (such as decimals, ticker...).
-    msg->tokenLookup1 = context->token_received;
-
-    msg->result = ETH_PLUGIN_RESULT_OK;
+    switch (context->selectorIndex) {
+        case SWAP_EXACT_ETH_FOR_TOKENS:
+            if (memcmp(msg->address, context->beneficiary, ADDRESS_LENGTH) != 0) {
+                msg->numScreens += 1;
+            }
+            msg->tokenLookup1 = context->token_received;
+            break;
+        case CLAIM_REGULAR_ACCOUNT:
+            break;
+        case CLAIM_MULTI_SIGNATURE_ACCOUNT:
+            break;
+        default:
+            msg->result = ETH_PLUGIN_RESULT_ERROR;
+            return;
+    }
 }
